requirements.txt
Flask
Flasgger
Flask-Limiter
openai
python-dotenv
fpdf2
python-docx
tensorflow
tensorflow-hub
scikit-learn
nltk
numpy
pdfplumber